+++ src/client/components/tm_urls_thead/tm_urls_thead.js
@@ -14,7 +14,6 @@
  */
 
 import _ from 'lodash';
-import { strict as assert } from 'node:assert';
 
 import { Forms } from 'meteor/pwix:forms';
 import { Logger } from 'meteor/pwix:logger';