+++ ChangeLog.md
@@ -29,6 +29,7 @@
     - Define new 'checkByTenant()' common function
     - Review async code from Records.s.upsert() using await instead of Promises
     - Prevent too many autorun's when calling setForm()
+    - Generalized URLs and emails have a single validity status per row
 
 ### 1.5.0